+++ src/ifdata_bcb/providers/ifdata/cadastro/__init__.py
@@ -1,4 +1,5 @@
 from ifdata_bcb.providers.ifdata.cadastro.collector import IFDATACadastroCollector
 from ifdata_bcb.providers.ifdata.cadastro.explorer import CadastroExplorer
+from ifdata_bcb.providers.ifdata.cadastro.search import CadastroSearch
 
-__all__ = ["IFDATACadastroCollector", "CadastroExplorer"]
+__all__ = ["IFDATACadastroCollector", "CadastroExplorer", "CadastroSearch"]